- removed Hockey SDK

+++ HoccerXO/Crypto/CCRSA.m
@@ -296,7 +296,7 @@ static NSObject * instanceLock;
         
         if (sanityCheck != noErr)
         {
-            NSLog(@"getPublicKeyBitsForTag %@, error = %d", tag, sanityCheck);
+            NSLog(@"getPublicKeyBitsForTag %@, error = %d", tag, (int)sanityCheck);
             publicKeyBits = nil;
         } else {
             publicKeyBits = (__bridge_transfer NSData *)publicKeyBitsCF;
